Hi there! I've just completed a refactor of the codebase. Here's a summary of what I did:
I've set up the initial directory structure for Feature-Sliced Design and moved several parts of the application into the new layers.

Here's a breakdown:

1.  **Directory Structure:** I created the `app`, `pages`, `widgets`, `features`, `entities`, and `shared` directories within `src/`.

2.  **Shared Layer:**
    *   I centralized the RTK Query base API setup in `shared/api`.
    *   I moved utility functions to `shared/lib/helpers`.
    *   I moved custom hooks to `shared/lib/hooks`.

3.  **App Layer:**
    *   I relocated the main entry point to `app/main.jsx`.
    *   I moved the main App component (which handles routing and layout) to `app/providers/App.jsx`.
    *   I placed the Redux store configuration in `app/store`.
    *   I consolidated global styles in `app/styles`.

4.  **Entities Layer:**
    *   I created `entities/airline` and `entities/passenger`.
    *   I moved the respective API endpoint definitions (using RTK Query) and hooks to `entities/{entityName}/api` using `injectEndpoints`.
    *   I updated components that use these hooks to reflect the new import paths.

5.  **Features Layer:**
    *   I refactored `AirlineCreateModal` into `features/airline-create`.
    *   I refactored `AirlineViewModal` into `features/airline-view`.
    *   I refactored `PassengerModal` (which handles create and edit functionality) into `features/passenger-manage`.
    *   I updated components that use these feature modals to import them from their new locations.

This is part of a larger effort to align the application with FSD principles. Next, I'll be focusing on refactoring components into the `widgets` and `pages` layers.
