+++ src/app/main.jsx
@@ -4,8 +4,8 @@ import ReactDOM from 'react-dom';
 import { Provider } from "react-redux"
 import { store } from './store'
 // Components
-import './index.css';
-import App from './App';
+import './styles/index.css';
+import App from './providers/App';
 
 ReactDOM.render(
     <Provider store={store}>
