+++ src/app/store/index.js
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit'
 import { setupListeners } from '@reduxjs/toolkit/query'
-import { generalApi } from './API'
+import { generalApi } from '../shared/api'
 
 export const store = configureStore({
     reducer: {
