+++ src/entities/airline/model/index.js
@@ -1,0 +1,1 @@
+// This file is intentionally left empty for now.
