+++ src/features/airline-create/index.js
@@ -1,0 +1,1 @@
+export { default as AirlineCreateModal } from './ui/AirlineCreateModal';
