+++ src/features/airline-view/index.js
@@ -1,0 +1,1 @@
+export { default as AirlineViewModal } from './ui/AirlineViewModal';
