+++ src/features/passenger-manage/index.js
@@ -1,0 +1,1 @@
+export { default as PassengerModal } from './ui/PassengerModal';
