+++ src/shared/lib/hooks/useObserver.js
